# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlancBleu API",
    "version": "1.3.0",
    "description": "Plateforme de gestion de transports sanitaires non urgents (Nice 06).\n\n**Authentification** : la plupart des routes exigent soit le cookie httpOnly `bb_access` (émis par /api/auth/login, prioritaire côté browser), soit le header `Authorization: Bearer <jwt>` (SPA / scripts). Les routes service-to-service (training-data, model/retrain) utilisent `X-Service-Token`.",
    "contact": {
      "name": "BlancBleu — Support",
      "email": "support@blancbleu.fr"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via POST /api/auth/login (utilisable en header Authorization)."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_access",
        "description": "Cookie httpOnly émis par /api/auth/login. Utilisé en priorité par les clients browser."
      },
      "serviceTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token",
        "description": "Token partagé Node ↔ Python pour les routes service-to-service (training-data, model/retrain, model/status)."
      }
    },
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "numeroPatient": {
            "type": "string",
            "example": "PAT-20260524-0001"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "dateNaissance": {
            "type": "string",
            "format": "date"
          },
          "telephone": {
            "type": "string"
          },
          "mobilite": {
            "type": "string",
            "enum": ["ASSIS", "FAUTEUIL_ROULANT", "ALLONGE", "CIVIERE"]
          },
          "oxygene": {
            "type": "boolean"
          },
          "brancardage": {
            "type": "boolean"
          },
          "accompagnateur": {
            "type": "boolean"
          },
          "source": {
            "type": "string",
            "enum": ["web", "app_mobile", "papier"]
          }
        }
      },
      "Facture": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "numero": {
            "type": "string",
            "example": "FAC-2026-0042"
          },
          "transportId": {
            "type": "string"
          },
          "montantTotal": {
            "type": "number"
          },
          "montantCPAM": {
            "type": "number"
          },
          "montantPatient": {
            "type": "number"
          },
          "statut": {
            "type": "string",
            "enum": [
              "brouillon",
              "emise",
              "en_attente",
              "payee",
              "annulee",
              "payment_failed",
              "remboursee",
              "partiellement_remboursee",
              "en_retard"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": ["UNPAID", "PENDING", "SUCCEEDED", "FAILED", "REFUNDED", "PARTIALLY_REFUNDED"]
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Token manquant, invalide ou expiré",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Rôle non autorisé pour cette action",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation des données échouée (corps de requête invalide)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/factures": {
      "get": {
        "tags": ["Factures"],
        "summary": "Liste paginée des factures",
        "description": "Filtres optionnels statut, paymentStatus, période d'émission, patient.",
        "parameters": [
          {
            "in": "query",
            "name": "statut",
            "schema": {
              "type": "string",
              "enum": [
                "brouillon",
                "emise",
                "en_attente",
                "payee",
                "annulee",
                "payment_failed",
                "remboursee",
                "en_retard"
              ]
            }
          },
          {
            "in": "query",
            "name": "paymentStatus",
            "schema": {
              "type": "string",
              "enum": ["UNPAID", "PENDING", "SUCCEEDED", "FAILED", "REFUNDED"]
            }
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de factures",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Facture"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/patients": {
      "get": {
        "tags": ["Patients"],
        "summary": "Liste paginée des patients (soft-deleted exclus)",
        "description": "Filtres optionnels nom, source, actif. Champs santé (`antecedents`, `allergies`) non remontés sauf opt-in via select.",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Recherche partielle nom/prénom/numeroPatient"
          },
          {
            "in": "query",
            "name": "source",
            "schema": {
              "type": "string",
              "enum": ["web", "app_mobile", "papier"]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Patient"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Patients"],
        "summary": "Créer un dossier patient",
        "description": "Numéro `PAT-YYYYMMDD-XXXX` attribué atomiquement. Le `numeroSecu` est\nchiffré AES-256-GCM et indexé via hash HMAC pour la recherche. RBAC :\nadmin / superviseur / dispatcher.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nom"],
                "properties": {
                  "nom": {
                    "type": "string"
                  },
                  "prenom": {
                    "type": "string"
                  },
                  "dateNaissance": {
                    "type": "string",
                    "format": "date"
                  },
                  "genre": {
                    "type": "string",
                    "enum": ["M", "F", "autre"]
                  },
                  "telephone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "numeroSecu": {
                    "type": "string",
                    "description": "Chiffré AES côté serveur"
                  },
                  "mobilite": {
                    "type": "string",
                    "enum": ["ASSIS", "FAUTEUIL_ROULANT", "ALLONGE", "CIVIERE"]
                  },
                  "oxygene": {
                    "type": "boolean"
                  },
                  "brancardage": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/transports": {
      "get": {
        "tags": ["Transports"],
        "summary": "Liste paginée des transports",
        "description": "Filtres combinables sur statut, période, véhicule, chauffeur, patient et type. Tri par dateTransport décroissant.",
        "parameters": [
          {
            "in": "query",
            "name": "statut",
            "schema": {
              "type": "string"
            },
            "description": "Filtrer par statut (REQUESTED, ASSIGNED, COMPLETED, …)"
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de transports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transport"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Transports"],
        "summary": "Créer un transport sanitaire",
        "description": "Crée un transport en statut `REQUESTED`. Géocodage automatique des\nadresses (BAN), création du dossier Patient si nécessaire, enqueue\ndu job auto-dispatch BullMQ. Le numéro `TRS-YYYYMMDD-XXXX` est\nattribué atomiquement (Counter Mongo).\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patient",
                  "typeTransport",
                  "motif",
                  "dateTransport",
                  "heureRDV",
                  "adresseDepart",
                  "adresseDestination"
                ],
                "properties": {
                  "patient": {
                    "$ref": "#/components/schemas/PatientSubdoc"
                  },
                  "typeTransport": {
                    "type": "string",
                    "enum": ["VSL", "TPMR", "AMBULANCE"]
                  },
                  "motif": {
                    "type": "string",
                    "example": "Dialyse"
                  },
                  "dateTransport": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "heureRDV": {
                    "type": "string",
                    "example": "09:30"
                  },
                  "allerRetour": {
                    "type": "boolean",
                    "default": false
                  },
                  "adresseDepart": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "adresseDestination": {
                    "$ref": "#/components/schemas/Adresse"
                  },
                  "tauxPriseEnCharge": {
                    "type": "number",
                    "default": 65
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transport créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/vehicles": {
      "get": {
        "tags": ["Vehicles"],
        "summary": "Liste paginée des véhicules",
        "description": "Filtres optionnels statut/type/disponible. Populate chauffeur assigné + transport en cours.",
        "parameters": [
          {
            "in": "query",
            "name": "statut",
            "schema": {
              "type": "string",
              "enum": ["Disponible", "En service", "Maintenance", "Hors service"]
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": ["VSL", "AMBULANCE", "TPMR"]
            }
          },
          {
            "in": "query",
            "name": "disponible",
            "schema": {
              "type": "boolean"
            },
            "description": "Raccourci équivalent à statut=Disponible"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de véhicules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vehicle"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Vehicles"],
        "summary": "Créer un véhicule (admin / superviseur)",
        "description": "L'immatriculation doit être unique. Le statut par défaut est Disponible.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["immatriculation", "nom", "type"],
                "properties": {
                  "immatriculation": {
                    "type": "string",
                    "example": "AB-123-CD"
                  },
                  "nom": {
                    "type": "string",
                    "example": "VSL-Niçois 01"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["VSL", "AMBULANCE", "TPMR"]
                  },
                  "statut": {
                    "type": "string",
                    "enum": ["Disponible", "En service", "Maintenance", "Hors service"],
                    "default": "Disponible"
                  },
                  "equipements": {
                    "type": "object",
                    "properties": {
                      "oxygene": {
                        "type": "boolean"
                      },
                      "fauteuilRampe": {
                        "type": "boolean"
                      },
                      "brancard": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Véhicule créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}